# Conflicts:
#	.vs/CleanBuildRose/v15/.suo
#	.vs/CleanBuildRose/v15/Browse.VC.db
#	rose/game.cpp

+++ rose/HealthBar.h
@@ -7,7 +7,7 @@ class HealthBar
 public:
 	const float maxHealth = 100;//This is just an idea for the max health
 	sf::Vector2f healthBarSize;
-	int currentHealth;
+	float currentHealth;
 	float currentHealthPercentage;
 
 	HealthBar(sf::RenderWindow &window);
